requirements.txt
python-dotenv==0.21.1
aiogram~=2.24
requests==2.28.2
beautifulsoup4==4.11.2
